requirements.txt
groq
python-dotenv
pyttsx3
jiwer
numpy
tqdm
